+++ usuariosflask/openapi_server/controllers/usuario_controller.py
@@ -1,19 +1,22 @@
-from openapi_server.models.usuario import Usuario  # noqa: E501
-from openapi_server.models.usuario_update import UsuarioUpdate  # noqa: E501
-from openapi_server import db
+# Se importa el fichero de configuración de los microservicios
+import os, sys, requests
+app_path = os.path.abspath(os.path.join(os.path.dirname(__file__), '../../..'))
+sys.path.append(app_path)
 
-# Importa la app de Flask
-from flask import request, jsonify
+from global_config import ContenidosConfig, UsuariosConfig
+
 from werkzeug.security import generate_password_hash, check_password_hash
 
-from openapi_server.models.usuario import Usuario
+from openapi_server import app, db
+from flask import request, jsonify
 
+from openapi_server.models.usuario import Usuario
 from openapi_server.models.usuario_db import UsuarioDB
+from openapi_server.models.usuario_update import UsuarioUpdate  # noqa: E501
+
 from openapi_server.models.perfil_db import PerfilDB
 from openapi_server.models.dispositivo_db import DispositivoDB
 from openapi_server.models.dispositivos_usuario_db import DispositivosUsuarioDB
-
-from openapi_server import app
 
 @app.route('/actualizar_usuario/<int:user_id>', methods=['PUT'])
 def actualizar_usuario(user_id):  # noqa: E501
@@ -138,14 +141,26 @@ def crear_usuario():  # noqa: E501
                 db.session.add(dispositivos_usuario_db)
 
         # Añadimos un perfil por defecto al usuario
-        perfiles_db = PerfilDB(nombre=usuario_db.nombre, user_id=usuario_db.user_id)
-        print(f"Datos del perfil_db:{perfiles_db.nombre}, {perfiles_db.user_id}")
+        # perfiles_db = PerfilDB(nombre=usuario_db.nombre, user_id=usuario_db.user_id)
+        # print(f"Datos del perfil_db:{perfiles_db.nombre}, {perfiles_db.user_id}")
 
-        db.session.add(perfiles_db)
+        # Se crear un perfil por defecto para el usuario
+        payload = {
+            "user_id": usuario_db.user_id,
+            "nombre": usuario_db.nombre,
+            "foto_perfil": 'netflux_rojo.png',
+            "preferencias_contenido": {},
+        }
+        request_perfil = requests.post(f"{UsuariosConfig.USUARIOS_BASE_URL}/usuario/{usuario_db.user_id}/perfiles", json=payload)
+
+        if request_perfil.status_code != 201:
+            return jsonify({"message": "Error al crear el perfil por defecto del usuario"}), 500
+
+        # db.session.add(perfiles_db)
         # db.session.update(usuario_db)
+
         db.session.commit()
         return jsonify({"message": "Usuario creado con éxito", "status": "success"}), 201
-
 
 @app.route('/eliminar_usuario/<user_id>', methods=['DELETE'])
 def eliminar_usuario(user_id):  # noqa: E501
@@ -158,18 +173,20 @@ def eliminar_usuario(user_id):  # noqa: E501
 
     :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
     """
-
     usuario = UsuarioDB.query.filter_by(user_id=user_id).first()
-
 
     if usuario is None:
         return jsonify({"message": "El usuario no existe", "status": "error"}), 404
     else:
+        request_borrar_perfiles = requests.delete(f"{UsuariosConfig.USUARIOS_BASE_URL}/usuario/{user_id}/perfiles")
+
+        if request_borrar_perfiles.status_code != 200:
+            return jsonify({"message": "Error al borrar los perfiles del usuario"}), 500
+
         db.session.delete(usuario)
         db.session.commit()
         return jsonify({"message": "Usuario eliminado con éxito", "status": "success"}), 200
     
-
 @app.route('/listar_usuarios', methods=['GET'])
 def listar_usuarios():  # noqa: E501
     """Listar todos los usuarios
@@ -179,11 +196,9 @@ def listar_usuarios():  # noqa: E501
 
     :rtype: Union[List[Usuario], Tuple[List[Usuario], int], Tuple[List[Usuario], int, Dict[str, str]]
     """
-    
     list_usuarios_db = UsuarioDB.query.all()
     list_usuarios_api = [usuario.to_api_model() for usuario in list_usuarios_db]
     return jsonify([usuario.serialize() for usuario in list_usuarios_api]), 200
-
 
 @app.route('/usuario/<user_id>', methods=['GET'])
 def obtener_usuario(user_id):  # noqa: E501
@@ -196,7 +211,6 @@ def obtener_usuario(user_id):  # noqa: E501
 
     :rtype: Union[Usuario, Tuple[Usuario, int], Tuple[Usuario, int, Dict[str, str]]
     """
-    
     usuario_db = UsuarioDB.query.filter_by(user_id=user_id).first()
     if usuario_db is None:
         return jsonify({"message": "El usuario no existe", "status": "error"}), 404
